Reworked how aggregations work
Now it embeds a full table into the cell of the result, and pass along a
`{:pass, {:op...}}` as the expression.

This opens the ideos of using nested SELECTS in the same way, just embed
the result as required, but more thought is required.

Works at the executor level, now necesary to be able to plan it.