Faster JOINs using maps
The new strategy creates a map for second table comparison key, and for
each
value on the first table of the comparison it generates the given rows.

This changes the algorithm complexity from N^2 to NlogN. And it is much
faster.

The old algorithm stays to more complex comparisons (not just
table1.field1 =
table2.field2) and very small tables.