Improved do not repeat :with test and fixes
It was failing on multiple withs statements, now at the test, and works

It was adding the how to execute :with at the context and at the 
queries, so
they were actually calling again the parsed expression. Now it is not 
added to
the context never and just added to the parse struct at the last moment, 
only on
the with statement (not on nested ones).